+++ apps/status-page/src/app/(status-page)/[domain]/layout.tsx
@@ -1,3 +1,4 @@
+import { defaultMetadata, ogMetadata, twitterMetadata } from "@/app/metadata";
 import {
   FloatingButton,
   StatusPageProvider,
@@ -5,6 +6,8 @@ import {
 import { ThemeProvider } from "@/components/theme-provider";
 import { Toaster } from "@/components/ui/sonner";
 import { HydrateClient, getQueryClient, trpc } from "@/lib/trpc/server";
+import type { Metadata } from "next";
+import { notFound } from "next/navigation";
 import { z } from "zod";
 
 export const schema = z.object({
@@ -54,4 +57,49 @@ export default async function Layout({
       </ThemeProvider>
     </HydrateClient>
   );
+}
+
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ domain: string }>;
+}): Promise<Metadata> {
+  const queryClient = getQueryClient();
+  const { domain } = await params;
+  const page = await queryClient.fetchQuery(
+    trpc.statusPage.get.queryOptions({ slug: domain }),
+  );
+
+  if (!page) return notFound();
+
+  return {
+    ...defaultMetadata,
+    title: {
+      template: `%s | ${page.title}`,
+      default: page?.title,
+    },
+    description: page?.description,
+    icons: page?.icon,
+    alternates: {
+      canonical: page?.customDomain
+        ? `https://${page.customDomain}`
+        : `https://${page.slug}.openstatus.dev`,
+    },
+    twitter: {
+      ...twitterMetadata,
+      images: [
+        `/api/og/page?slug=${page?.slug}&passwordProtected=${page?.passwordProtected}`,
+      ],
+      title: page?.title,
+      description: page?.description,
+    },
+    openGraph: {
+      ...ogMetadata,
+      images: [
+        `/api/og/page?slug=${page?.slug}&passwordProtected=${page?.passwordProtected}`,
+      ],
+      title: page?.title,
+      description: page?.description,
+    },
+  };
 }
